Validates all images by running a Java app (#59)
Validates all images by running a Java app

+++ validate-image.sh
@@ -42,19 +42,37 @@ if [[ $? -ne 0 ]]; then
     exit 1
 fi
 
-# Validate the image
-if [[ "${distro}" == "distroless" ]]; then
-    java_version=$(docker run --rm $image -version 2>&1 | head -n 1 | awk -F '"' '{print $2}')
-else
-    java_version=$(docker run --rm $image /bin/bash -c "source \$JAVA_HOME/release && echo \$JAVA_VERSION")
+# Validate the image if expectedversion is set (not blank)
+if [[ ! -z "$expectedversion" ]]; then
+  if [[ "${distro}" == "distroless" ]]; then
+      java_version=$(docker run --rm $image -version 2>&1 | head -n 1 | awk -F '"' '{print $2}')
+  else
+      java_version=$(docker run --rm $image /bin/bash -c "source \$JAVA_HOME/release && echo \$JAVA_VERSION")
+  fi
+
+  java_version=${java_version//[$'\t\r\n']}
+  java_version=${java_version%%*( )}
+
+  if [[ "$java_version" == "$expectedversion" ]]; then
+      echo "::notice title=Validation succeeded ($jdkversion-$distro)::Image '${image}' contains expected JDK version: ${expectedversion}"
+  else
+      echo "::error title=Wrong minor JDK version ($jdkversion-$distro)::Image '${image}' contains unexpected JDK version: ${java_version}. Expected: ${expectedversion}."
+      exit 1
+  fi
 fi
 
-java_version=${java_version//[$'\t\r\n']}
-java_version=${java_version%%*( )}
+# Test running a Java app
+dockerfile="./docker/test-only/Dockerfile.testapp"
+if [[ "${distro}" == "distroless" ]]; then
+  dockerfile=${dockerfile}"distroless"
+fi
 
-if [[ "$java_version" == "$expectedversion" ]]; then
-    echo "::notice title=Validation succeeded ($jdkversion-$distro)::Image '${image}' contains expected JDK version: ${expectedversion}"
+docker build --build-arg IMGTOTEST=$image -t testapprunner -f $dockerfile ./docker/test-only/
+test_output=$(docker run --rm testapprunner)
+
+if [[ "${test_output}" =~ "Hello World" ]]; then
+    echo "::notice title=Test of sample app SUCCEEDED ($jdkversion-$distro)::Image '${image}' is ABLE to run a sample Java app."
 else
-    echo "::error title=Wrong minor JDK version ($jdkversion-$distro)::Container image '${image}' contains unexpected JDK version: ${java_version}. Expected: ${expectedversion}."
+    echo "::error title=Test of sample app FAILED ($jdkversion-$distro)::Image '${image}' CANNOT run a sample Java app."
     exit 1
 fi
